Finish moving the scenario tests to integration tests
