+++ pipeline/merger/src/test/scala/weco/pipeline/merger/MergerIntegrationTest.scala
@@ -3,7 +3,8 @@ package weco.pipeline.merger
 import org.scalatest.GivenWhenThen
 import org.scalatest.featurespec.AnyFeatureSpec
 import org.scalatest.matchers.should.Matchers
-import weco.catalogue.internal_model.work.Format
+import weco.catalogue.internal_model.identifiers.IdentifierType
+import weco.catalogue.internal_model.work.{CollectionPath, Format, InvisibilityReason}
 import weco.catalogue.internal_model.work.generators.SourceWorkGenerators
 import weco.fixtures.TimeAssertions
 import weco.pipeline.matcher.generators.MergeCandidateGenerators
@@ -53,9 +54,9 @@ class MergerIntegrationTest
       val miro2 = miroIdentifiedWork()
       val miro3 = miroIdentifiedWork()
       val sierra = sierraPhysicalIdentifiedWork()
-        .mergeCandidates(List(miro1, miro2, miro3).map(createMiroMergeCandidateFor))
+        .mergeCandidates(List(miro1, miro2, miro3).map(createMiroSierraMergeCandidateFor))
 
-      When("the works are merged")
+      When("the works are processed by the matcher/merger")
       processWorks(sierra, miro1, miro2, miro3)
 
       Then("the Miro works are redirected to the Sierra work")
@@ -81,9 +82,9 @@ class MergerIntegrationTest
       Given("a Sierra work and a Miro work")
       val miro = miroIdentifiedWork()
       val sierra = sierraPhysicalIdentifiedWork()
-        .mergeCandidates(List(createMiroMergeCandidateFor(miro)))
+        .mergeCandidates(List(createMiroSierraMergeCandidateFor(miro)))
 
-      When("the works are merged")
+      When("the works are processed by the matcher/merger")
       processWorks(sierra, miro)
 
       Then("the Miro work is redirected to the Sierra work")
@@ -109,7 +110,7 @@ class MergerIntegrationTest
       val mets = metsIdentifiedWork()
         .mergeCandidates(List(createMetsMergeCandidateFor(sierraPicture)))
 
-      When("the works are merged")
+      When("the works are processed by the matcher/merger")
       processWorks(sierraPicture, mets)
 
       Then("the METS work is redirected to the Sierra work")
@@ -135,7 +136,7 @@ class MergerIntegrationTest
       val mets = metsIdentifiedWork()
         .mergeCandidates(List(createMetsMergeCandidateFor(sierraEphemera)))
 
-      When("the works are merged")
+      When("the works are processed by the matcher/merger")
       processWorks(sierraEphemera, mets)
 
       Then("the METS work is redirected to the Sierra work")
@@ -149,6 +150,480 @@ class MergerIntegrationTest
         .getMerged(sierraEphemera)
         .data
         .imageData should contain only mets.singleImage
+    }
+  }
+
+  Scenario("An AIDS poster Sierra picture, a METS and a Miro are matched") {
+    withContext { implicit context =>
+      Given(
+        "a Sierra picture with digcode `digaids`, a METS work and a Miro work"
+      )
+      val miro = miroIdentifiedWork()
+      val sierraDigaidsPicture = sierraIdentifiedWork()
+        // Multiple physical items would prevent a Miro redirect in any other case,
+        // but we still expect to see it for the digaids works as the Miro item is
+        // a known duplicate of the METS item.
+        .items(List(createIdentifiedPhysicalItem, createIdentifiedPhysicalItem))
+        .format(Format.Pictures)
+        .otherIdentifiers(List(createDigcodeIdentifier("digaids")))
+        .mergeCandidates(List(createMiroSierraMergeCandidateFor(miro)))
+      val mets = metsIdentifiedWork()
+        .mergeCandidates(List(createMetsMergeCandidateFor(sierraDigaidsPicture)))
+
+      When("the works are processed by the matcher/merger")
+      processWorks(sierraDigaidsPicture, mets, miro)
+
+      Then("the METS work and the Miro work are redirected to the Sierra work")
+      context.getMerged(mets) should beRedirectedTo(sierraDigaidsPicture)
+      context.getMerged(miro) should beRedirectedTo(sierraDigaidsPicture)
+
+      And("the Sierra work contains only the METS images")
+      context
+        .getMerged(sierraDigaidsPicture)
+        .data
+        .imageData should contain only mets.singleImage
+    }
+  }
+
+  Scenario("A physical and a digital Sierra work are matched") {
+    withContext { implicit context =>
+      Given("a pair of a physical Sierra work and a digital Sierra work")
+      val (digitalSierra, physicalSierra) = sierraIdentifiedWorkPair()
+
+      When("the works are processed by the matcher/merger")
+      processWorks(physicalSierra, digitalSierra)
+
+      Then("the digital work is redirected to the physical work")
+      context.getMerged(digitalSierra) should beRedirectedTo(physicalSierra)
+
+      And("the physical work contains the digitised work's identifiers")
+      val physicalIdentifiers = context.getMerged(physicalSierra).identifiers
+      physicalIdentifiers should contain allElementsOf digitalSierra.identifiers
+    }
+  }
+
+  Scenario("Audiovisual Sierra works are not merged") {
+    withContext { implicit context =>
+      Given("a physical Sierra AV work and its digitised counterpart")
+      val digitisedVideo =
+        sierraDigitalIdentifiedWork().format(Format.EVideos)
+
+      val physicalVideo =
+        sierraPhysicalIdentifiedWork()
+          .format(Format.Videos)
+          .mergeCandidates(
+            List(createSierraPairMergeCandidateFor(digitisedVideo)))
+
+      When("the works are processed by the matcher/merger")
+      processWorks(physicalVideo, digitisedVideo)
+
+      Then("both original works remain visible")
+      context.getMerged(physicalVideo) should beVisible
+      context.getMerged(digitisedVideo) should beVisible
+    }
+  }
+
+  Scenario("A Calm work and a Sierra work are matched") {
+    withContext { implicit context =>
+      Given("a Sierra work and a Calm work")
+      val calm = calmIdentifiedWork()
+      val sierra = sierraPhysicalIdentifiedWork()
+        .mergeCandidates(List(createCalmMergeCandidateFor(calm)))
+
+      When("the works are processed by the matcher/merger")
+      processWorks(sierra, calm)
+
+      Then("the Sierra work is redirected to the Calm work")
+      context.getMerged(sierra) should beRedirectedTo(calm)
+
+      And("the Calm work contains the Sierra item ID")
+      val calmItem = context.getMerged(calm).data.items.head
+      calmItem.id shouldBe sierra.data.items.head.id
+    }
+  }
+
+  Scenario("A Calm work, a Sierra work, and a Miro work are matched") {
+    withContext { implicit context =>
+      Given("A Calm work, a Sierra work and a Miro work")
+      val calm = calmIdentifiedWork()
+      val miro = miroIdentifiedWork()
+      val sierra = sierraPhysicalIdentifiedWork()
+        .mergeCandidates(
+          List(
+            createMiroSierraMergeCandidateFor(miro),
+            createCalmMergeCandidateFor(calm)
+          )
+        )
+
+      When("the works are processed by the matcher/merger")
+      processWorks(sierra, calm, miro)
+
+      Then("the Sierra work is redirected to the Calm work")
+      context.getMerged(sierra) should beRedirectedTo(calm)
+
+      And("the Miro work is redirected to the Calm work")
+      context.getMerged(miro) should beRedirectedTo(calm)
+
+      And("the Calm work contains the Miro location")
+      context.getMerged(calm).data.items.flatMap(_.locations) should
+        contain(miro.data.items.head.locations.head)
+      And("the Calm work contains the Miro image")
+      context.getMerged(calm).data.imageData should contain(miro.singleImage)
+    }
+  }
+
+  Scenario("A Calm work, a Sierra picture work, and a METS work are matched") {
+    withContext { implicit context =>
+      Given("A Calm work, a Sierra picture work and a METS work")
+      val calm = calmIdentifiedWork()
+      val sierraPicture = sierraIdentifiedWork()
+        .items(List(createIdentifiedPhysicalItem))
+        .format(Format.Pictures)
+        .mergeCandidates(List(createCalmMergeCandidateFor(calm)))
+      val mets = metsIdentifiedWork()
+        .mergeCandidates(List(createMetsMergeCandidateFor(sierraPicture)))
+
+      When("the works are processed by the matcher/merger")
+      processWorks(sierraPicture, calm, mets)
+
+      Then("the Sierra work is redirected to the Calm work")
+      context.getMerged(sierraPicture) should beRedirectedTo(calm)
+
+      And("the METS work is redirected to the Calm work")
+      context.getMerged(mets) should beRedirectedTo(calm)
+
+      And("the Calm work contains the METS location")
+      context.getMerged(calm).data.items.flatMap(_.locations) should
+        contain(mets.data.items.head.locations.head)
+      And("the Calm work contains the METS image")
+      context.getMerged(calm).data.imageData should contain(mets.singleImage)
+    }
+  }
+
+  Scenario("A Calm work and multiple Miro works are matched") {
+    withContext { implicit context =>
+      Given("A Calm work and 2 Miro works")
+      val miro1 = miroIdentifiedWork()
+      val miro2 = miroIdentifiedWork()
+      val calm = calmIdentifiedWork()
+        .mergeCandidates(
+          List(
+            createCalmMiroMergeCandidateFor(miro1),
+            createCalmMiroMergeCandidateFor(miro2),
+          )
+        )
+
+      When("the works are processed by the matcher/merger")
+      processWorks(calm, miro1, miro2)
+
+      Then("the Miro works are redirected to the Calm work")
+      context.getMerged(miro1) should beRedirectedTo(calm)
+      context.getMerged(miro2) should beRedirectedTo(calm)
+
+      And("the Calm work contains the miro items")
+      context
+        .getMerged(calm)
+        .data
+        .items should contain allElementsOf miro1.data.items
+      context
+        .getMerged(calm)
+        .data
+        .items should contain allElementsOf miro2.data.items
+
+      And("the Calm work contains the Miro images")
+      context.getMerged(calm).data.imageData should contain(miro1.singleImage)
+      context.getMerged(calm).data.imageData should contain(miro2.singleImage)
+    }
+  }
+
+  Scenario("A digitised video with Sierra physical records and e-bibs") {
+    withContext { implicit context =>
+      // This test case is based on a real example of four related works that
+      // were being merged incorrectly.  In particular, the METS work (and associated
+      // IIIF manifest) was being merged into the physical video formats, not the
+      // more detailed e-bib that it should have been attached to.
+      //
+      // See https://wellcome.slack.com/archives/C3TQSF63C/p1615474389063800
+      Given("a Sierra physical record, an e-bib, and a METS work")
+      val workWithPhysicalVideoFormats =
+        sierraIdentifiedWork()
+          .title("A work with physical video formats, e.g. DVD or digibeta")
+          .format(Format.Film)
+          .items(List(createIdentifiedPhysicalItem))
+
+      val workForEbib =
+        sierraIdentifiedWork()
+          .title("A work for an e-bib")
+          .format(Format.Videos)
+          .mergeCandidates(
+            List(
+              createSierraPairMergeCandidateFor(workWithPhysicalVideoFormats))
+          )
+
+      val workForMets =
+        identifiedWork(sourceIdentifier = createMetsSourceIdentifier)
+          .title("A METS work")
+          .mergeCandidates(List(createMetsMergeCandidateFor(workForEbib)))
+          .items(List(createDigitalItem))
+          .invisible()
+
+      When("the works are processed by the matcher/merger")
+      processWorks(workWithPhysicalVideoFormats, workForEbib, workForMets)
+
+      Then("the METS work is redirected to the Sierra e-bib")
+      context.getMerged(workForMets) should beRedirectedTo(workForEbib)
+
+      And("the Sierra e-bib gets the items from the METS work")
+      context.getMerged(workForEbib).data.items shouldBe workForMets.data.items
+
+      And("the Sierra physical work is unaffected")
+      context.getMerged(workWithPhysicalVideoFormats).data shouldBe workWithPhysicalVideoFormats.data
+      context.getMerged(workWithPhysicalVideoFormats).state should beSimilarTo(workWithPhysicalVideoFormats.state)
+    }
+  }
+
+  Scenario("A Tei and a Sierra digital and a sierra physical work are merged") {
+    withContext { implicit context =>
+      Given("a Tei, a Sierra physical record and a Sierra digital record")
+      val (digitalSierra, physicalSierra) = sierraIdentifiedWorkPair()
+      val teiWork = teiIdentifiedWork()
+        .mergeCandidates(List(createTeiBnumberMergeCandidateFor(physicalSierra)))
+
+      When("the works are processed by the matcher/merger")
+      processWorks(digitalSierra, physicalSierra, teiWork)
+
+      Then("the Sierra works are redirected to the tei")
+      context.getMerged(digitalSierra) should beRedirectedTo(teiWork)
+      context.getMerged(physicalSierra) should beRedirectedTo(teiWork)
+
+      And("the tei work has the Sierra works' items")
+      context
+        .getMerged(teiWork)
+        .data
+        .items should contain allElementsOf digitalSierra.data.items
+      context
+        .getMerged(teiWork)
+        .data
+        .items should contain allElementsOf physicalSierra.data.items
+
+      And("the tei work has the Sierra works' identifiers")
+      context
+        .getMerged(teiWork)
+        .data
+        .otherIdentifiers should contain allElementsOf physicalSierra.data.otherIdentifiers
+        .filter(_.identifierType == IdentifierType.SierraIdentifier)
+      context
+        .getMerged(teiWork)
+        .data
+        .otherIdentifiers should contain allElementsOf digitalSierra.data.otherIdentifiers
+        .filter(_.identifierType == IdentifierType.SierraIdentifier)
+    }
+  }
+
+  Scenario(
+    "A Tei with internal works and a Sierra digital and a sierra physical work are merged") {
+    withContext { implicit context =>
+      Given("a Tei, a Sierra physical record and a Sierra digital record")
+      val (digitalSierra, physicalSierra) = sierraIdentifiedWorkPair()
+      val firstInternalWork =
+        teiIdentifiedWork().collectionPath(CollectionPath("1"))
+      val secondInternalWork =
+        teiIdentifiedWork().collectionPath(CollectionPath("2"))
+
+      val teiWork = teiIdentifiedWork()
+        .title("A tei work")
+        .internalWorks(List(firstInternalWork, secondInternalWork))
+        .mergeCandidates(List(createTeiBnumberMergeCandidateFor(physicalSierra)))
+
+      When("the works are processed by the matcher/merger")
+      processWorks(digitalSierra, physicalSierra, teiWork)
+
+      Then("the Sierra works are redirected to the tei")
+      context.getMerged(digitalSierra) should beRedirectedTo(teiWork)
+      context.getMerged(physicalSierra) should beRedirectedTo(teiWork)
+
+      And("the tei work has the Sierra works' items")
+      context
+        .getMerged(teiWork)
+        .data
+        .items should contain allElementsOf digitalSierra.data.items
+      context
+        .getMerged(teiWork)
+        .data
+        .items should contain allElementsOf physicalSierra.data.items
+
+      And("the tei work has the Sierra works' identifiers")
+      context
+        .getMerged(teiWork)
+        .data
+        .otherIdentifiers should contain allElementsOf physicalSierra.data.otherIdentifiers
+        .filter(_.identifierType == IdentifierType.SierraIdentifier)
+      context
+        .getMerged(teiWork)
+        .data
+        .otherIdentifiers should contain allElementsOf digitalSierra.data.otherIdentifiers
+        .filter(_.identifierType == IdentifierType.SierraIdentifier)
+
+      And("the internal tei works are returned")
+      context.getMerged(firstInternalWork) should beVisible
+      context.getMerged(secondInternalWork) should beVisible
+
+      And("the tei internal works contain the sierra item")
+      context
+        .getMerged(firstInternalWork)
+        .data
+        .items should contain allElementsOf physicalSierra.data.items
+      context
+        .getMerged(secondInternalWork)
+        .data
+        .items should contain allElementsOf physicalSierra.data.items
+
+      And("the tei internal works retain their collectionsPath")
+      context
+        .getMerged(firstInternalWork)
+        .data
+        .collectionPath shouldBe firstInternalWork.data.collectionPath
+      context
+        .getMerged(secondInternalWork)
+        .data
+        .collectionPath shouldBe secondInternalWork.data.collectionPath
+    }
+  }
+
+  Scenario("A Tei work passes through unchanged") {
+    withContext { implicit context =>
+      Given("a Tei")
+      val internalWork1 = teiIdentifiedWork()
+      val internalWork2 = teiIdentifiedWork()
+      val teiWork = teiIdentifiedWork()
+        .title("A tei work")
+        .internalWorks(List(internalWork1, internalWork2))
+
+      When("the tei work is merged")
+      processWork(teiWork)
+
+      Then("the tei work should be a TEI work")
+      val mergedWork = context.getMerged(teiWork)
+      mergedWork.data shouldBe teiWork.data
+      mergedWork.state should beSimilarTo(teiWork.state)
+
+      And("the the tei inner works should be returned")
+      Seq(internalWork1, internalWork2).foreach { w =>
+        val expectedWork = updateInternalWork(w, teiWork)
+
+        context.getMerged(w).data shouldBe expectedWork.data
+        context.getMerged(w).state should beSimilarTo(expectedWork.state)
+      }
+    }
+  }
+
+  Scenario(
+    "CollectionPath is prepended to internal tei works if the work is not merged") {
+    withContext { implicit context =>
+      Given("a Tei")
+      val internalWork1 =
+        teiIdentifiedWork().collectionPath(CollectionPath("id/1"))
+      val internalWork2 =
+        teiIdentifiedWork().collectionPath(CollectionPath("id/2"))
+      val teiWork = teiIdentifiedWork()
+        .title("A tei work")
+        .internalWorks(List(internalWork1, internalWork2))
+        .collectionPath(CollectionPath("id"))
+
+      When("the work is processed by the matcher/merger")
+      processWork(teiWork)
+
+      Then("the tei work should be a TEI work")
+      val expectedInternalWork1 = updateInternalWork(internalWork1, teiWork)
+        .collectionPath(CollectionPath("id/1"))
+      val expectedInternalWork2 = updateInternalWork(internalWork2, teiWork)
+        .collectionPath(CollectionPath("id/2"))
+
+      context.getMerged(internalWork1).data shouldBe expectedInternalWork1.data
+      context.getMerged(internalWork1).state should beSimilarTo(expectedInternalWork1.state)
+
+      context.getMerged(internalWork2).data shouldBe expectedInternalWork2.data
+      context.getMerged(internalWork2).state should beSimilarTo(expectedInternalWork2.state)
+
+      val expectedTeiWork = teiWork.internalWorks(List(expectedInternalWork1, expectedInternalWork2))
+      context.getMerged(teiWork).data shouldBe expectedTeiWork.data
+      context.getMerged(teiWork).state should beSimilarTo(expectedTeiWork.state)
+    }
+  }
+
+  Scenario("A TEI work, a Calm work, a Sierra work and a METS work") {
+    withContext { implicit context =>
+      Given("four works")
+      val calmWork =
+        calmIdentifiedWork()
+          .otherIdentifiers(
+            List(
+              createSourceIdentifierWith(
+                identifierType = IdentifierType.CalmRefNo),
+              createSourceIdentifierWith(
+                identifierType = IdentifierType.CalmAltRefNo),
+            )
+          )
+          .items(List(createCalmItem))
+
+      val sierraWork =
+        sierraIdentifiedWork()
+          .otherIdentifiers(
+            List(
+              createSourceIdentifierWith(
+                identifierType = IdentifierType.SierraIdentifier),
+              createSourceIdentifierWith(
+                identifierType = IdentifierType.WellcomeDigcode),
+            )
+          )
+          .items(List(createIdentifiedPhysicalItem))
+          .mergeCandidates(List(createCalmMergeCandidateFor(calmWork)))
+
+      val teiWork = teiIdentifiedWork()
+        .mergeCandidates(List(createTeiBnumberMergeCandidateFor(sierraWork)))
+
+      val metsWork =
+        metsIdentifiedWork()
+          .thumbnail(createDigitalLocation)
+          .items(List(createDigitalItem))
+          .mergeCandidates(List(createMetsMergeCandidateFor(sierraWork)))
+          .invisible(invisibilityReasons =
+            List(InvisibilityReason.MetsWorksAreNotVisible))
+
+      When("the works are processed by the matcher/merger")
+      processWorks(teiWork, sierraWork, metsWork, calmWork)
+
+      Then("Everything should be redirected to the TEI work")
+      context.getMerged(sierraWork) should beRedirectedTo(teiWork)
+      context.getMerged(metsWork) should beRedirectedTo(teiWork)
+      context.getMerged(calmWork) should beRedirectedTo(teiWork)
+
+      And("the TEI work gets all the CALM and Sierra identifiers")
+      val teiMergedIdentifiers =
+        context
+          .getMerged(teiWork)
+          .data
+          .otherIdentifiers
+
+      teiMergedIdentifiers should contain allElementsOf calmWork.data.otherIdentifiers :+ calmWork.state.sourceIdentifier
+      teiMergedIdentifiers should contain allElementsOf sierraWork.data.otherIdentifiers :+ sierraWork.state.sourceIdentifier
+
+      And("it has no METS identifier")
+      teiMergedIdentifiers.filter(_.identifierType == IdentifierType.METS) shouldBe empty
+
+      And("it only has two items (one physical, one digital)")
+      val teiItems =
+        context
+          .getMerged(teiWork)
+          .data
+          .items
+
+      teiItems should contain allElementsOf sierraWork.data.items
+      teiItems should contain allElementsOf metsWork.data.items
+      teiItems should contain noElementsOf calmWork.data.items
+
+      And("it gets the METS thumbnail")
+      context.getMerged(teiWork).data.thumbnail shouldBe metsWork.data.thumbnail
     }
   }
 }
